backend/requirements.txt
# Web Framework
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5

# Document Processing
python-docx>=1.0.0
pypdf2>=3.0.0

# Use simpler tokenizers rather than sentence-transformers for now
# We'll install sentence-transformers separately
transformers>=4.0.0

# Vector Database - use simpler version without complex build dependencies
faiss-cpu>=1.7.0

# API clients
openai>=1.0.0
requests>=2.20.0
huggingface-hub>=0.13.0

# Numpy - use precompiled version
numpy>=1.20.0